Refactor YouTube script to match layout of other publisher scripts
